Split large Cypress index test out
