+++ cypress/e2e/court-cases-index/roles.cy.ts
@@ -1,0 +1,178 @@
+import {
+  loginAndGoToUrl,
+  confirmCaseDisplayed,
+  confirmCaseNotDisplayed,
+  confirmReasonDisplayed,
+  confirmReasonNotDisplayed
+} from "../../support/helpers"
+
+describe("Shows relevant information to a user's role", () => {
+  const mixedReasonCases = [
+    {
+      orgForPoliceFilter: "011111",
+      errorId: 0,
+      exceptions: [],
+      triggers: []
+    },
+    {
+      orgForPoliceFilter: "011111",
+      errorId: 1,
+      exceptions: [
+        {
+          code: "HO100310",
+          field: "ds:OffenceReasonSequence"
+        },
+        {
+          code: "HO100301",
+          field: "ds:OffenceReasonSequence"
+        }
+      ],
+      triggers: []
+    },
+    {
+      orgForPoliceFilter: "011111",
+      errorId: 2,
+      exceptions: [],
+      triggers: [
+        {
+          triggerId: 0,
+          triggerCode: "TRPR0010",
+          status: "Unresolved",
+          createdAt: new Date()
+        },
+        {
+          triggerId: 1,
+          triggerCode: "TRPR0011",
+          status: "Unresolved",
+          createdAt: new Date()
+        }
+      ]
+    },
+    {
+      orgForPoliceFilter: "011111",
+      errorId: 3,
+      exceptions: [
+        {
+          code: "HO100310",
+          field: "ds:OffenceReasonSequence"
+        },
+        {
+          code: "HO100301",
+          field: "ds:OffenceReasonSequence"
+        }
+      ],
+      triggers: [
+        {
+          triggerId: 2,
+          triggerCode: "TRPR0010",
+          status: "Unresolved",
+          createdAt: new Date()
+        },
+        {
+          triggerId: 3,
+          triggerCode: "TRPR0011",
+          status: "Unresolved",
+          createdAt: new Date()
+        }
+      ]
+    }
+  ]
+
+  beforeEach(() => {
+    cy.task(
+      "insertCourtCasesWithFields",
+      mixedReasonCases.map((courtCase) => {
+        return {
+          errorId: courtCase.errorId,
+          orgForPoliceFilter: courtCase.orgForPoliceFilter
+        }
+      })
+    )
+    mixedReasonCases.forEach((courtCase) => {
+      courtCase.exceptions?.forEach((exception) => {
+        cy.task("insertException", {
+          caseId: courtCase.errorId,
+          exceptionCode: exception.code,
+          errorReport: `${exception.code}||${exception.field}`
+        })
+      })
+      if (courtCase.triggers && courtCase.triggers.length > 0) {
+        cy.task("insertTriggers", { caseId: courtCase.errorId, triggers: courtCase.triggers })
+      }
+    })
+  })
+
+  it("Shouldn't show cases to a user with no groups", () => {
+    loginAndGoToUrl("nogroups@example.com")
+
+    cy.findByText("There are no court cases to show").should("exist")
+  })
+
+  it("Should only show cases with triggers to a trigger handler", () => {
+    loginAndGoToUrl("triggerhandler@example.com")
+
+    confirmCaseDisplayed("Case00002")
+    confirmCaseDisplayed("Case00003")
+
+    confirmCaseNotDisplayed("Case00000")
+    confirmCaseNotDisplayed("Case00001")
+  })
+
+  it("Should only show cases with exceptions to an exception handler", () => {
+    loginAndGoToUrl("exceptionhandler@example.com")
+
+    confirmCaseDisplayed("Case00001")
+    confirmCaseDisplayed("Case00003")
+
+    confirmCaseNotDisplayed("Case00000")
+    confirmCaseNotDisplayed("Case00002")
+  })
+
+  it("Should show all cases to a general handler", () => {
+    loginAndGoToUrl("generalhandler@example.com")
+
+    confirmCaseDisplayed("Case00000")
+    confirmCaseDisplayed("Case00001")
+    confirmCaseDisplayed("Case00002")
+    confirmCaseDisplayed("Case00003")
+  })
+
+  it("Should show all cases to a supervisor", () => {
+    loginAndGoToUrl("supervisor@example.com")
+
+    confirmCaseDisplayed("Case00000")
+    confirmCaseDisplayed("Case00001")
+    confirmCaseDisplayed("Case00002")
+    confirmCaseDisplayed("Case00003")
+  })
+
+  it("Should only show triggers in the reason column to a trigger handler", () => {
+    loginAndGoToUrl("triggerhandler@example.com")
+
+    confirmReasonDisplayed("TRPR0010")
+    confirmReasonDisplayed("TRPR0011")
+
+    confirmReasonNotDisplayed("HO100301")
+    confirmReasonNotDisplayed("HO100310")
+  })
+
+  it("Should only show exceptions in the reason column to an exception handler", () => {
+    loginAndGoToUrl("exceptionhandler@example.com")
+
+    confirmReasonNotDisplayed("TRPR0010")
+    confirmReasonNotDisplayed("TRPR0011")
+
+    confirmReasonDisplayed("HO100301")
+    confirmReasonDisplayed("HO100310")
+  })
+
+  it("Should show both triggers and exceptions in the reason column to a general handler", () => {
+    loginAndGoToUrl("generalhandler@example.com")
+
+    confirmReasonDisplayed("TRPR0010")
+    confirmReasonDisplayed("TRPR0011")
+
+    confirmReasonDisplayed("HO100301")
+    confirmReasonDisplayed("HO100310")
+  })
+})
